separate VotingProgress
