+++ app/components/indicators/index.js
@@ -14,7 +14,7 @@ export { default as LoaderBarBottom } from "./LoaderBarBottom";
 export { default as NoStats } from "./NoStats";
 export { default as NoTransactions } from "./NoTransactions";
 export { default as StepIndicator } from "./StepIndicator/StepIndicator";
-export { default as VotingProgress } from "./VotingProgress";
+export { default as VotingProgress } from "./VotingProgress/VotingProgress";
 export { default as NoProposals } from "./NoProposals";
 export { default as PoliteiaLoading } from "./PoliteiaLoading";
 export { default as ProgressRing } from "./ProgressRing";
